backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "newsapi-python>=0.2.7",
    "orjson>=3.11.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "tavily-python>=0.7.10",
    "uvicorn[standard]>=0.35.0",
]
